+++ gaztec/src/components/home/home.tsx
@@ -1,16 +1,12 @@
 import { css } from "@emotion/react";
 import { ContractComponent } from "../contract/contract";
 import { SidebarComponent } from "../sidebar/sidebar";
-import { createContext, useState } from "react";
-import {
-  type PXE,
-  type AccountWalletWithSecretKey,
-  Contract,
-  AztecNode,
-  AztecAddress,
-} from "@aztec/aztec.js";
-import { type WalletDB } from "../../utils/storage";
-import { ContractFunctionInteractionTx } from "../../utils/txs";
+import { useState } from "react";
+import { AztecContext } from "../../aztecEnv";
+import NoSleep from "nosleep.js";
+import { LogPanel } from "../logPanel/logPanel";
+import logoURL from "../../assets/Aztec_logo.png";
+import { Box, Drawer } from "@mui/material";
 
 const layout = css({
   display: "flex",
@@ -18,47 +14,30 @@ const layout = css({
   height: "100%",
 });
 
-export const AztecContext = createContext<{
-  pxe: PXE | null;
-  nodeURL: string;
-  node: AztecNode;
-  wallet: AccountWalletWithSecretKey | null;
-  isPXEInitialized: boolean;
-  walletDB: WalletDB | null;
-  currentContractAddress: AztecAddress;
-  currentContract: Contract;
-  currentTx: ContractFunctionInteractionTx;
-  setWalletDB: (walletDB: WalletDB) => void;
-  setPXEInitialized: (isPXEInitialized: boolean) => void;
-  setWallet: (wallet: AccountWalletWithSecretKey) => void;
-  setAztecNode: (node: AztecNode) => void;
-  setPXE: (pxe: PXE) => void;
-  setNodeURL: (nodeURL: string) => void;
-  setCurrentTx: (currentTx: ContractFunctionInteractionTx) => void;
-  setCurrentContract: (currentContract: Contract) => void;
-  setCurrentContractAddress: (currentContractAddress: AztecAddress) => void;
-}>({
-  pxe: null,
-  nodeURL: "",
-  node: null,
-  wallet: null,
-  isPXEInitialized: false,
-  walletDB: null,
-  currentContract: null,
-  currentContractAddress: null,
-  currentTx: null,
-  setWalletDB: (walletDB: WalletDB) => {},
-  setPXEInitialized: (isPXEInitialized: boolean) => {},
-  setWallet: (wallet: AccountWalletWithSecretKey) => {},
-  setNodeURL: (nodeURL: string) => {},
-  setPXE: (pxe: PXE) => {},
-  setAztecNode: (node: AztecNode) => {},
-  setCurrentTx: (currentTx: ContractFunctionInteractionTx) => {},
-  setCurrentContract: (currentContract: Contract) => {},
-  setCurrentContractAddress: (currentContractAddress: AztecAddress) => {},
+const logo = css({
+  width: "100%",
+  padding: "0.5rem",
 });
 
-export function Home() {
+const collapsedDrawer = css({
+  height: "100%",
+  width: "4rem",
+  backgroundColor: "var(--mui-palette-primary-light)",
+  overflow: "hidden",
+});
+
+const noSleep = new NoSleep();
+
+function enableNoSleep() {
+  noSleep.enable();
+  document.removeEventListener("touchstart", enableNoSleep, false);
+}
+
+// Enable wake lock.
+// (must be wrapped in a user input event handler e.g. a mouse or touch handler)
+document.addEventListener("touchstart", enableNoSleep, false);
+
+export default function Home() {
   const [pxe, setPXE] = useState(null);
   const [wallet, setWallet] = useState(null);
   const [nodeURL, setNodeURL] = useState("");
@@ -69,6 +48,9 @@ export function Home() {
   const [currentContract, setCurrentContract] = useState(null);
   const [currentTx, setCurrentTx] = useState(null);
   const [currentContractAddress, setCurrentContractAddress] = useState(null);
+  const [logs, setLogs] = useState([]);
+  const [logsOpen, setLogsOpen] = useState(false);
+  const [drawerOpen, setDrawerOpen] = useState(false);
 
   const AztecContextInitialValue = {
     pxe,
@@ -81,6 +63,12 @@ export function Home() {
     currentTx,
     node,
     currentContractAddress,
+    logs,
+    logsOpen,
+    drawerOpen,
+    setDrawerOpen,
+    setLogsOpen,
+    setLogs,
     setAztecNode,
     setCurrentTx,
     setWalletDB,
@@ -96,7 +84,23 @@ export function Home() {
   return (
     <div css={layout}>
       <AztecContext.Provider value={AztecContextInitialValue}>
-        <SidebarComponent />
+        <div css={collapsedDrawer} onClick={() => setDrawerOpen(!drawerOpen)}>
+          <img css={logo} src={logoURL} />
+        </div>
+        <Drawer
+          sx={{
+            "& .MuiDrawer-paper": {
+              height: "100%",
+              width: "340px",
+            },
+          }}
+          onClose={() => setDrawerOpen(false)}
+          variant="temporary"
+          open={drawerOpen}
+        >
+          <SidebarComponent />
+        </Drawer>
+        <LogPanel />
         <ContractComponent />
       </AztecContext.Provider>
     </div>