+++ spartan/terraform/deploy-metrics/variables.tf
@@ -1,7 +1,7 @@
 variable "GKE_CLUSTER_CONTEXT" {
   description = "GKE cluster context"
   type        = string
-  default     = "gke_testnet-440309_us-east4-a_spartan-gke"
+  default     = "gke_testnet-440309_us-west1-a_aztec-gke-private"
 }
 
 variable "RELEASE_NAME" {
@@ -12,4 +12,20 @@ variable "RELEASE_NAME" {
 variable "VALUES_FILE" {
   description = "Name of the values file to use for deployment"
   type        = string
+}
+
+variable "GRAFANA_DASHBOARD_PASSWORD" {
+  description = "Grafana dashboard password"
+  type        = string
+  sensitive   = true
+}
+
+variable "project" {
+  default = "testnet-440309"
+  type    = string
+}
+
+variable "region" {
+  default = "us-west1"
+  type    = string
 }