+++ yarn-project/end-to-end/src/e2e_outbox.test.ts
@@ -56,8 +56,8 @@ describe('E2E Outbox Tests', () => {
     ];
 
     const call = new BatchCall(wallets[0], [
-      contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content1, recipient1).request(),
-      contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content2, recipient2).request(),
+      await contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content1, recipient1).request(),
+      await contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content2, recipient2).request(),
     ]);
 
     // TODO (#5104): When able to guarantee multiple txs in a single block, make this populate a full tree. Right now we are
@@ -176,9 +176,9 @@ describe('E2E Outbox Tests', () => {
     ];
 
     const call0 = new BatchCall(wallets[0], [
-      contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content1, recipient1).request(),
-      contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content2, recipient2).request(),
-      contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content3, recipient3).request(),
+      await contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content1, recipient1).request(),
+      await contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content2, recipient2).request(),
+      await contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content3, recipient3).request(),
     ]);
 
     const call1 = contract.methods.create_l2_to_l1_message_arbitrary_recipient_private(content4, recipient4);