+++ yarn-project/end-to-end/src/e2e_synching.test.ts
@@ -48,23 +48,37 @@ import {
 import { createBlobSinkClient } from '@aztec/blob-sink/client';
 // eslint-disable-next-line no-restricted-imports
 import { L2Block, tryStop } from '@aztec/circuit-types';
-import { type AztecAddress } from '@aztec/circuits.js';
-import { getL1ContractsConfigEnvVars } from '@aztec/ethereum';
-import { Timer } from '@aztec/foundation/timer';
+import { type AztecAddress, EthAddress } from '@aztec/circuits.js';
+import { EpochCache } from '@aztec/epoch-cache';
+import {
+  GovernanceProposerContract,
+  L1TxUtilsWithBlobs,
+  RollupContract,
+  SlashingProposerContract,
+  getL1ContractsConfigEnvVars,
+} from '@aztec/ethereum';
+import { TestDateProvider, Timer } from '@aztec/foundation/timer';
 import { RollupAbi } from '@aztec/l1-artifacts';
 import { SchnorrHardcodedAccountContract } from '@aztec/noir-contracts.js/SchnorrHardcodedAccount';
 import { SpamContract } from '@aztec/noir-contracts.js/Spam';
 import { TokenContract } from '@aztec/noir-contracts.js/Token';
 import { type PXEService } from '@aztec/pxe';
-import { L1Publisher } from '@aztec/sequencer-client';
+import { SequencerPublisher } from '@aztec/sequencer-client';
 import { createWorldStateSynchronizer } from '@aztec/world-state';
 
 import * as fs from 'fs';
 import { getContract } from 'viem';
 
+import { DEFAULT_BLOB_SINK_PORT } from './fixtures/fixtures.js';
 import { addAccounts } from './fixtures/snapshot_manager.js';
 import { mintTokensToPrivate } from './fixtures/token_utils.js';
-import { type EndToEndContext, getPrivateKeyFromIndex, setup, setupPXEService } from './fixtures/utils.js';
+import {
+  type EndToEndContext,
+  createForwarderContract,
+  getPrivateKeyFromIndex,
+  setup,
+  setupPXEService,
+} from './fixtures/utils.js';
 
 const SALT = 420;
 const AZTEC_GENERATE_TEST_DATA = !!process.env.AZTEC_GENERATE_TEST_DATA;
@@ -230,10 +244,10 @@ class TestVariant {
       const txs = [];
       for (let i = 0; i < this.txCount; i++) {
         const batch = new BatchCall(this.wallets[i], [
-          this.spam.methods.spam(this.seed, 16, false).request(),
-          this.spam.methods.spam(this.seed + 16n, 16, false).request(),
-          this.spam.methods.spam(this.seed + 32n, 16, false).request(),
-          this.spam.methods.spam(this.seed + 48n, 15, true).request(),
+          await this.spam.methods.spam(this.seed, 16, false).request(),
+          await this.spam.methods.spam(this.seed + 16n, 16, false).request(),
+          await this.spam.methods.spam(this.seed + 32n, 16, false).request(),
+          await this.spam.methods.spam(this.seed + 48n, 15, true).request(),
         ]);
 
         this.seed += 100n;
@@ -331,7 +345,7 @@ describe('e2e_synching', () => {
 
       // Now we create all of our interesting blocks.
       // Alter the block requirements for the sequencer such that we ensure blocks sizes as desired.
-      sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
+      await sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
 
       // The setup will mint tokens (private and public)
       await variant.setup();
@@ -382,10 +396,34 @@ describe('e2e_synching', () => {
     await (sequencer as any).stop();
     await watcher?.stop();
 
-    const blobSinkClient = createBlobSinkClient(`http://localhost:${blobSink?.port ?? 5052}`);
+    const blobSinkClient = createBlobSinkClient({
+      blobSinkUrl: `http://localhost:${blobSink?.port ?? DEFAULT_BLOB_SINK_PORT}`,
+    });
 
     const sequencerPK: `0x${string}` = `0x${getPrivateKeyFromIndex(0)!.toString('hex')}`;
-    const publisher = new L1Publisher(
+
+    const l1TxUtils = new L1TxUtilsWithBlobs(
+      deployL1ContractsValues.publicClient,
+      deployL1ContractsValues.walletClient,
+      logger,
+      config,
+    );
+    const rollupAddress = deployL1ContractsValues.l1ContractAddresses.rollupAddress.toString();
+    const rollupContract = new RollupContract(deployL1ContractsValues.publicClient, rollupAddress);
+    const governanceProposerContract = new GovernanceProposerContract(
+      deployL1ContractsValues.publicClient,
+      config.l1Contracts.governanceProposerAddress.toString(),
+    );
+    const slashingProposerAddress = await rollupContract.getSlashingProposerAddress();
+    const slashingProposerContract = new SlashingProposerContract(
+      deployL1ContractsValues.publicClient,
+      slashingProposerAddress.toString(),
+    );
+    const forwarderContract = await createForwarderContract(config, sequencerPK, rollupAddress);
+    const epochCache = await EpochCache.create(config.l1Contracts.rollupAddress, config, {
+      dateProvider: new TestDateProvider(),
+    });
+    const publisher = new SequencerPublisher(
       {
         l1RpcUrl: config.l1RpcUrl,
         requiredConfirmations: 1,
@@ -396,8 +434,17 @@ describe('e2e_synching', () => {
         viemPollingIntervalMS: 100,
         ethereumSlotDuration: ETHEREUM_SLOT_DURATION,
         blobSinkUrl: `http://localhost:${blobSink?.port ?? 5052}`,
+        customForwarderContractAddress: EthAddress.ZERO,
       },
-      { blobSinkClient },
+      {
+        blobSinkClient,
+        l1TxUtils,
+        rollupContract,
+        forwarderContract,
+        governanceProposerContract,
+        slashingProposerContract,
+        epochCache,
+      },
     );
 
     const blocks = variant.loadBlocks();
@@ -411,7 +458,7 @@ describe('e2e_synching', () => {
         await cheatCodes.eth.mine();
       }
       // If it breaks here, first place you should look is the pruning.
-      await publisher.proposeL2Block(block);
+      await publisher.enqueueProposeL2Block(block);
     }
 
     await alternativeSync({ deployL1ContractsValues, cheatCodes, config, logger, pxe }, variant);
@@ -500,7 +547,9 @@ describe('e2e_synching', () => {
             await aztecNode.stop();
           }
 
-          const blobSinkClient = createBlobSinkClient(`http://localhost:${opts.blobSink?.port ?? 5052}`);
+          const blobSinkClient = createBlobSinkClient({
+            blobSinkUrl: `http://localhost:${opts.blobSink?.port ?? DEFAULT_BLOB_SINK_PORT}`,
+          });
           const archiver = await createArchiver(opts.config!, blobSinkClient, {
             blockUntilSync: true,
           });
@@ -514,7 +563,7 @@ describe('e2e_synching', () => {
           const assumeProvenThrough = pendingBlockNumber - 2n;
           await rollup.write.setAssumeProvenThroughBlockNumber([assumeProvenThrough]);
 
-          const timeliness = (await rollup.read.EPOCH_DURATION()) * 2n;
+          const timeliness = (await rollup.read.getEpochDuration()) * 2n;
           const blockLog = await rollup.read.getBlock([(await rollup.read.getProvenBlockNumber()) + 1n]);
           const timeJumpTo = await rollup.read.getTimestampForSlot([blockLog.slotNumber + timeliness]);
 
@@ -525,10 +574,10 @@ describe('e2e_synching', () => {
           const txHash = blockTip.body.txEffects[0].txHash;
 
           const contractClassIds = await archiver.getContractClassIds();
-          contracts.forEach(async c => {
+          for (const c of contracts) {
             expect(contractClassIds.includes(c.instance.contractClassId)).toBeTrue;
             expect(await archiver.getContract(c.address)).not.toBeUndefined;
-          });
+          }
 
           expect(await archiver.getTxEffect(txHash)).not.toBeUndefined;
           expect(await archiver.getPrivateLogs(blockTip.number, 1)).not.toEqual([]);
@@ -600,7 +649,7 @@ describe('e2e_synching', () => {
 
           const blockBeforePrune = await aztecNode.getBlockNumber();
 
-          const timeliness = (await rollup.read.EPOCH_DURATION()) * 2n;
+          const timeliness = (await rollup.read.getEpochDuration()) * 2n;
           const blockLog = await rollup.read.getBlock([(await rollup.read.getProvenBlockNumber()) + 1n]);
           const timeJumpTo = await rollup.read.getTimestampForSlot([blockLog.slotNumber + timeliness]);
 
@@ -626,7 +675,7 @@ describe('e2e_synching', () => {
 
           const blockBefore = await aztecNode.getBlock(await aztecNode.getBlockNumber());
 
-          sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
+          await sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
           const txs = await variant.createAndSendTxs();
           await Promise.all(txs.map(tx => tx.wait({ timeout: 1200 })));
 
@@ -660,7 +709,7 @@ describe('e2e_synching', () => {
           const pendingBlockNumber = await rollup.read.getPendingBlockNumber();
           await rollup.write.setAssumeProvenThroughBlockNumber([pendingBlockNumber - BigInt(variant.blockCount) / 2n]);
 
-          const timeliness = (await rollup.read.EPOCH_DURATION()) * 2n;
+          const timeliness = (await rollup.read.getEpochDuration()) * 2n;
           const blockLog = await rollup.read.getBlock([(await rollup.read.getProvenBlockNumber()) + 1n]);
           const timeJumpTo = await rollup.read.getTimestampForSlot([blockLog.slotNumber + timeliness]);
 
@@ -685,7 +734,7 @@ describe('e2e_synching', () => {
 
           const blockBefore = await aztecNode.getBlock(await aztecNode.getBlockNumber());
 
-          sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
+          await sequencer?.updateSequencerConfig({ minTxsPerBlock: variant.txCount, maxTxsPerBlock: variant.txCount });
           const txs = await variant.createAndSendTxs();
           await Promise.all(txs.map(tx => tx.wait({ timeout: 1200 })));