+++ yarn-project/merkle-tree/src/pedersen.ts
@@ -1,4 +1,4 @@
-import { pedersenHash } from '@aztec/foundation/crypto';
+import { pedersenHash } from '@aztec/foundation/crypto/sync';
 import { Fr } from '@aztec/foundation/fields';
 import { type Hasher } from '@aztec/types/interfaces';