+++ yarn-project/validator-client/src/metrics.ts
@@ -40,11 +40,11 @@ export class ValidatorMetrics {
     this.reExecutionTime.record(time);
   }
 
-  public recordFailedReexecution(proposal: BlockProposal) {
+  public async recordFailedReexecution(proposal: BlockProposal) {
     this.failedReexecutionCounter.add(1, {
       [Attributes.STATUS]: 'failed',
       [Attributes.BLOCK_NUMBER]: proposal.payload.header.globalVariables.blockNumber.toString(),
-      [Attributes.BLOCK_PROPOSER]: proposal.getSender()?.toString(),
+      [Attributes.BLOCK_PROPOSER]: (await proposal.getSender())?.toString(),
     });
   }
 }