+++ yarn-project/aztec/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [0.73.0](https://github.com/AztecProtocol/aztec-packages/compare/aztec-package-v0.72.1...aztec-package-v0.73.0) (2025-02-01)
+
+
+### Features
+
+* **archiver:** Read blobs from cl ([#11273](https://github.com/AztecProtocol/aztec-packages/issues/11273)) ([8cf4982](https://github.com/AztecProtocol/aztec-packages/commit/8cf4982175da82575087bb28590a3a71b88aaf6e))
+* Lazy wasm pt4 ([#11491](https://github.com/AztecProtocol/aztec-packages/issues/11491)) ([f553f58](https://github.com/AztecProtocol/aztec-packages/commit/f553f58ee5f405414a7d0c402c679ffe5c4d62ef))
+* Proposer forwarder contract ([#11403](https://github.com/AztecProtocol/aztec-packages/issues/11403)) ([76f25b6](https://github.com/AztecProtocol/aztec-packages/commit/76f25b6d674c658bc01a42267f4fb232a35d5551))
+* Validate L1 config against L1 on startup ([#11540](https://github.com/AztecProtocol/aztec-packages/issues/11540)) ([48b7ac4](https://github.com/AztecProtocol/aztec-packages/commit/48b7ac4dd823533843018e177e0cf2bd00d1bdba))
+
+
+### Miscellaneous
+
+* Enable no-misused-promises ([#11629](https://github.com/AztecProtocol/aztec-packages/issues/11629)) ([a420466](https://github.com/AztecProtocol/aztec-packages/commit/a420466cfb0a387e98a932be8982fe25f02b90db))
+
+## [0.72.1](https://github.com/AztecProtocol/aztec-packages/compare/aztec-package-v0.72.0...aztec-package-v0.72.1) (2025-01-24)
+
+
+### Miscellaneous
+
+* **aztec-package:** Synchronize aztec-packages versions
+
+## [0.72.0](https://github.com/AztecProtocol/aztec-packages/compare/aztec-package-v0.71.0...aztec-package-v0.72.0) (2025-01-24)
+
+
+### Features
+
+* Gaztec ([#11229](https://github.com/AztecProtocol/aztec-packages/issues/11229)) ([79f810d](https://github.com/AztecProtocol/aztec-packages/commit/79f810dc682d41154eb723e5bdf4c54c0681becb))
+* Lazy wasm pt. 2 ([#11410](https://github.com/AztecProtocol/aztec-packages/issues/11410)) ([01510f4](https://github.com/AztecProtocol/aztec-packages/commit/01510f45aa5d385a08584df674d9caf9522e6be2))
+* Lazy wasm pt3 ([#11435](https://github.com/AztecProtocol/aztec-packages/issues/11435)) ([7068d05](https://github.com/AztecProtocol/aztec-packages/commit/7068d055d91a6e81e6fbb670e17c77ee209a1a80))
+
+
+### Bug Fixes
+
+* Init fee juice contract in sandbox ([#11379](https://github.com/AztecProtocol/aztec-packages/issues/11379)) ([caab526](https://github.com/AztecProtocol/aztec-packages/commit/caab52671cfcf20b395a9e44a8768dc81d986cb5))
+* Use simulation to estimate gas used ([#11211](https://github.com/AztecProtocol/aztec-packages/issues/11211)) ([63776f0](https://github.com/AztecProtocol/aztec-packages/commit/63776f0d217fad800bf8a6c6144d6bb52844e629))
+
+
+### Miscellaneous
+
+* Trace propagation from json rpc client to server ([#11325](https://github.com/AztecProtocol/aztec-packages/issues/11325)) ([85ccc15](https://github.com/AztecProtocol/aztec-packages/commit/85ccc1512cd9b1c461660ad8127dae848fde1878))
+
 ## [0.71.0](https://github.com/AztecProtocol/aztec-packages/compare/aztec-package-v0.70.0...aztec-package-v0.71.0) (2025-01-17)